fava/requirements.txt
attrs==21.4.0
Babel==2.10.1
beancount==2.3.5
beautifulsoup4==4.11.1
bottle==0.12.19
cachetools==5.1.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Babel==2.0.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
grpcio==1.50.0
grpcio-tools==1.50.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
markdown2==2.4.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
ply==3.11
protobuf==4.21.9
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2022.1
requests==2.27.1
rsa==4.8
simplejson==3.17.6
soupsieve==2.3.2.post1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2

